+++ src/molara/structure/io/importer.py
@@ -15,6 +15,7 @@ from molara.structure.io.importer_crystal import PoscarImporter, PymatgenImporte
 from molara.structure.molecule import Molecule
 from molara.structure.molecules import Molecules
 from molara.structure.mos import Mos
+from molara.util.exceptions import FileFormatError, FileImporterError
 
 if TYPE_CHECKING:
     from collections.abc import Mapping
@@ -31,14 +32,6 @@ if TYPE_CHECKING:
 __copyright__ = "Copyright 2024, Molara"
 
 bohr_to_angstrom = 5.29177210903e-1
-
-
-class FileImporterError(Exception):
-    """Base class for errors occurring when loading molecules from file."""
-
-
-class FileFormatError(FileImporterError):
-    """Raised when the file format is wrong or unsupported."""
 
 
 class MoleculesImporter(ABC):
