+++ src/molara/util/exceptions.py
@@ -6,6 +6,8 @@ from __future__ import annotations
 class FileImporterError(Exception):
     """base class for errors occurring when loading molecules from file."""
 
+class FileExporterError(Exception):
+    """Base class for errors occurring when loading molecules from file."""
 
-class FileFormatError(FileImporterError):
+class FileFormatError(ValueError):
     """raised when the file format is wrong or unsupported."""
